refactor: extract test to configurator_test
